Implemented most of the packages, and also most of the methods in client

+++ Turnbased-Game/Models/Client/IParticipant.cs
@@ -13,7 +13,7 @@ public interface IParticipant
     public event Action BadRequest; 
 
     public void ListAvailableLobbies();
-    public void JoinLobby(int lobbyId);
+    public void JoinLobby(byte lobbyId);
     public void DisconnectLobby();
     public void IsReady();
     public void IsNotReady();
